requirements.txt
pyyaml
openai
tenacity
mcp
fastmcp
faiss-cpu
fastapi==0.124.0
click>=8.1.8,<8.3
uvicorn
websockets
wsproto
pydantic==2.12.5
requests
pytest
ddgs
beautifulsoup4
matplotlib
networkx
cartopy
pandas>=2.3.3
openpyxl>=3.1.2
numpy>=2.3.5
seaborn>=0.13.2
google-genai>=1.52.0
chardet>=5.2.0
pygame>=2.6.1
filelock>=3.20.1
markdown>=3.10
xhtml2pdf>=0.2.17
bcrypt
gunicorn
psycopg2-binary
PyJWT
